+++ app/api/servers/delete-api/route.ts
@@ -4,11 +4,10 @@ import { NextResponse } from "next/server";
 
 export async function DELETE(
     req: Request,
-    { params }: { params: { serverId: string } }
   ) {
     try {
       const profile = await currentUserProfile();
-  
+      const { serverId } = await req.json();
       if (!profile) {
         throw new Error("Unauthorized");
       }
@@ -21,7 +20,7 @@ export async function DELETE(
       // });
       const server = await db.server.updateMany({
         where: {
-          id: params.serverId,
+          id: serverId,
           userProfileId: profile.id,
         },
         data: {